Make DIFFUSION_ATTENTION_QUANT generic; Cast to FP8 from FP32.
- Currently DIFFUSION_ATTENTION_QUANT is parsed to AttnQuantSpec, which
is only used by FLASHINFER_ATTN and TRTLLM_ATTN. But the link between
DIFFUSION_ATTENTION_QUANT<>AttnQuantSpec is generic. No reason to block
this parsing with backend-specific if-conditions.
- Casting FP16->FP8 could be risky. Cast from FP32 instead.

+++ vllm_omni/diffusion/attention/backends/flashinfer_attn.py
@@ -206,7 +206,7 @@ class FlashInferAttentionImpl(AttentionImpl):
     ) -> tuple[torch.Tensor, float]:
         if torch.finfo(to_dtype).bits == 8:
             scale_tensor = tensor.abs().amax().float().clamp_min(1e-6) / torch.finfo(to_dtype).max
-            tensor = (tensor * torch.reciprocal(scale_tensor).to(tensor.dtype)).to(to_dtype)
+            tensor = (tensor.float() * torch.reciprocal(scale_tensor)).to(to_dtype)
             return tensor, FlashInferAttentionImpl._extract_scalar(scale_tensor)
         return tensor.to(to_dtype), 1.0
